+++ zagar_protocol/src/protocol/model/Cell.java
@@ -8,10 +8,10 @@ public final class Cell {
     private final int cellId;
     private final int playerId;
     private final float size;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Cell(int cellId, int playerId, float size, int x, int y) {
+    public Cell(int cellId, int playerId, float size, double x, double y) {
         this.cellId = cellId;
         this.playerId = playerId;
         this.size = size;
@@ -23,11 +23,11 @@ public final class Cell {
         return playerId;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
